Cargo clippy

+++ node/src/bootstrap/bootstrap_server.rs
@@ -58,7 +58,7 @@ pub struct BootstrapServer {
 
 impl BootstrapServer {
     /** Maximum number of blocks to send in a single response, cannot be higher than capacity of a single `asc_pull_ack` message */
-    pub const MAX_BLOCKS: u8 = BlocksAckPayload::MAX_BLOCKS as u8;
+    pub const MAX_BLOCKS: u8 = BlocksAckPayload::MAX_BLOCKS;
     pub const MAX_FRONTIERS: usize = AscPullAck::MAX_FRONTIERS;
 
     pub(crate) fn new(
